Issue Fix Area 1 Data Model bug fixes resolved 4 issues identified
1.1 resolved a major bug where in the holdings the symbol was made the id and primary key which caused only one copy of an symbol to exists through out the system, resolved it by adding an long id field

1.2 batch find prices had a small issue of(incorrect max date) but this is nto an problem but an intended solution as all prices have same max dates but i did update the batchfind to send a map of symbol,price to not have any sorting later one and get accurate symbol data

1.3 again same as 1.2 but since i do final sorting of the values it is as i intended and works till i get a more better solution

1.4 when storing the percentage of the allocations or sector limits we want to store a double but instead used Integer whihc made the point precision go away and data to be vanished so changed all strategy related dto and service to move from Integer to Double

+++ src/main/java/majorproject/maf/dto/request/StrategyRequestDto.java
@@ -17,7 +17,7 @@ public class StrategyRequestDto {
     private String goal;
     private Integer timeHorizonMonths;
     private RiskProfile riskProfile;
-    private Map<String, Integer> targetAllocation; // or Map<String, BigDecimal>
-    private Map<String, Integer> sectorLimits;
+    private Map<String, Double> targetAllocation; // or Map<String, BigDecimal>
+    private Map<String, Double> sectorLimits;
     private RebalancingFrequency rebalancingFrequency;
 }
